convert to int

+++ module-4-2/module-4-2.py
@@ -3,4 +3,4 @@ def input_func(input):
     start , end = map(int, input.split(":"))
     n = end - start + 1
 
-    return (n / 2) * (2*start + (n-1))
+    return int((n / 2) * (2*start + (n-1)))
